requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
Django==2.2.9
django-ckeditor==5.9.0
django-debug-toolbar==3.2.2
django-js-asset==1.2.2
easy-thumbnails==2.7
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
Pillow==7.1.2
pkg_resources==0.0.0
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
pytest-django==3.8.0
python-dotenv==0.19.2
pytz==2020.1
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
wcwidth==0.1.9
zipp==3.1.0
